+++ modules/webhook/eventbridge/dispatcher.tf
@@ -5,20 +5,19 @@ resource "aws_cloudwatch_event_rule" "workflow_job" {
 
   event_pattern = <<EOF
 {
-  "detail-type": [
-    "workflow_job"
-  ]
+  "detail-type": ["workflow_job"],
+  "detail": {
+    "action": ["queued"]
+  }
 }
 EOF
 }
 
-
-resource "aws_cloudwatch_event_target" "github_welcome" {
+resource "aws_cloudwatch_event_target" "dispatcher" {
   arn            = aws_lambda_function.dispatcher.arn
   rule           = aws_cloudwatch_event_rule.workflow_job.name
   event_bus_name = aws_cloudwatch_event_bus.main.name
 }
-
 
 resource "aws_lambda_function" "dispatcher" {
   s3_bucket         = var.config.lambda_s3_bucket != null ? var.config.lambda_s3_bucket : null
@@ -45,8 +44,8 @@ resource "aws_lambda_function" "dispatcher" {
         POWERTOOLS_TRACER_CAPTURE_ERROR          = var.config.tracing_config.capture_error
         # Parameters required for lambda configuration
         PARAMETER_RUNNER_MATCHER_CONFIG_PATH = var.config.ssm_parameter_runner_matcher_config.name
+        PARAMETER_RUNNER_MATCHER_VERSION     = var.config.ssm_parameter_runner_matcher_config.version # enforce cold start after Changes in SSM parameter
         REPOSITORY_ALLOW_LIST                = jsonencode(var.config.repository_white_list)
-        SQS_WORKFLOW_JOB_QUEUE               = try(var.config.sqs_workflow_job_queue.id, null)
       } : k => v if v != null
     }
   }
@@ -66,10 +65,6 @@ resource "aws_lambda_function" "dispatcher" {
     content {
       mode = var.config.tracing_config.mode
     }
-  }
-
-  lifecycle {
-    replace_triggered_by = [null_resource.ssm_parameter_runner_matcher_config, null_resource.github_app_parameters]
   }
 }
 
@@ -142,14 +137,4 @@ resource "aws_iam_role_policy" "dispatcher_xray" {
   name   = "xray-policy"
   policy = data.aws_iam_policy_document.lambda_xray[0].json
   role   = aws_iam_role.dispatcher_lambda.name
-}
-
-resource "aws_iam_role_policy" "dispatcher_workflow_job_sqs" {
-  count = var.config.sqs_workflow_job_queue != null ? 1 : 0
-  name  = "publish-workflow-job-sqs-policy"
-  role  = aws_iam_role.dispatcher_lambda.name
-
-  policy = templatefile("${path.module}/../policies/lambda-publish-sqs-policy.json", {
-    sqs_resource_arns = jsonencode([var.config.sqs_workflow_job_queue.arn])
-  })
 }